refactor: Break out farm fetch into own helper (#1332)
* refactor: Break out farm fetch into own helper

* refactor: Updated fetchFarm method

Moved merging of farm config up a level to fetch farms
